# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RunPipelineBody {
  /** Kaggle dataset identifier (e.g. "titanic" or "username/dataset-name") */
  datasetName: string;
  /**
   * Optional target column name. Auto-detected if not provided.
   * @nullable
   */
  targetColumn?: string | null;
  /**
   * Optional Kaggle username for API auth
   * @nullable
   */
  kaggleUsername?: string | null;
  /**
   * Optional Kaggle API key for auth
   * @nullable
   */
  kaggleKey?: string | null;
}

export type PipelineJobStatus =
  (typeof PipelineJobStatus)[keyof typeof PipelineJobStatus];

export const PipelineJobStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface PipelineJob {
  jobId: string;
  status: PipelineJobStatus;
  /**
   * Current stage description
   * @nullable
   */
  stage?: string | null;
  datasetName: string;
  /** @nullable */
  targetColumn?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  error?: string | null;
}

export type PipelineResultProblemType =
  (typeof PipelineResultProblemType)[keyof typeof PipelineResultProblemType];

export const PipelineResultProblemType = {
  classification: "classification",
  regression: "regression",
} as const;

export interface ModelScore {
  modelName: string;
  score: number;
  trainingTime: number;
}

export interface FeatureImportance {
  feature: string;
  importance: number;
}

export interface DatasetSummary {
  rows: number;
  columns: number;
  columnNames: string[];
  missingValues: number;
  duplicatesRemoved: number;
  categoricalColumns: string[];
  numericColumns: string[];
}

export interface PipelineResult {
  jobId: string;
  datasetName: string;
  targetColumn: string;
  problemType: PipelineResultProblemType;
  bestModel: string;
  bestScore: number;
  scoreMetric: string;
  modelScores: ModelScore[];
  featureImportance: FeatureImportance[];
  /** Feature names used during training (for building prediction form) */
  featureNames: string[];
  datasetSummary: DatasetSummary;
  /** @nullable */
  modelPath?: string | null;
}

/**
 * Map of feature name to string value entered by user
 */
export type PredictBodyFeatures = { [key: string]: string };

export interface PredictBody {
  /** Map of feature name to string value entered by user */
  features: PredictBodyFeatures;
}

export type PredictResultProblemType =
  (typeof PredictResultProblemType)[keyof typeof PredictResultProblemType];

export const PredictResultProblemType = {
  classification: "classification",
  regression: "regression",
} as const;

export interface PredictResult {
  /** The model's predicted output as a string */
  predictedValue: string;
  /** Top 5 most important features */
  topFeatures: FeatureImportance[];
  /** Human-readable explanation of prediction */
  explanation: string;
  /**
   * High, Medium, or Low confidence level
   * @nullable
   */
  confidence?: string | null;
  /**
   * Metric value used to compute confidence
   * @nullable
   */
  confidenceDetail?: string | null;
  problemType: PredictResultProblemType;
}

export interface KaggleDataset {
  ref: string;
  title: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  lastUpdated?: string | null;
  /** @nullable */
  downloadCount?: number | null;
}

export interface ErrorResponse {
  error: string;
}

export type SearchKaggleDatasetsParams = {
  query: string;
};